Many timer changes and BMS request added ADC rate code. More testing is needed, but seems-to-work. Further experiments with EXTI to detect end of ADC conversion still shows SDO lines does not go high. Tried jumper PC4 to PB4 (PI MISO) to assure EXTI code interrupted, but to no avail

+++ bmsadbms1818/Ourtasks/CanCommTask.h
@@ -23,7 +23,7 @@
  * @param	: taskpriority = Task priority (just as it says!)
  * @return	: CanCommHandle
  * *************************************************************************/
-void CanComm_qreq(uint32_t reqcode, uint32_t setfets, struct CANRCVBUF* pcan);
+void CanComm_qreq(uint8_t reqcode, uint32_t setfets, struct CANRCVBUF* pcan);
 /*	@brief	: Queue request to BMSTask.c
  *  @param  : reqcode = see BMSTask.h 
  *  @param  : setfets = bits to set discharge fets (if so commanded)
